Remove customised method name on each controller for gem entry
Originally seemed like a neat idea to create a method name
constructed from the controllers class name. eg. it would create
`book_parsms` when called from the following `BooksController`.

However after some discussion this seems a but cumbersome and
possibly error prone. eg currently. it creates a method called
`application_params` on the `ApplicationController` class
which could be troublesome ahead if not expecting it.

From now on you simply call `tidy_params` it it will return the
appropriate whitelist for whichever controller it's called from